Use modified mention regex from client, more tests
Uses the same mention regex as chat-gui does to highlight messages now,
with the exception that `mentionedusername.` will only count as a
mention if it *isn't* immediately followed by a non-whitespace
character. For example, mentioning destiny,

- `destiny` -> counts as a mention
- `destiny.` -> counts as a mention
- `destiny. hi` -> counts as a mention
- `destiny.gg` -> **does not** count as a mention
- `destiny. gg` -> counts count as a mention

The implication of this is that urls won't also count as mentions,
specifically to deal with the common case of someone typing `destiny.gg`
and getting muted even while not trying to mention destiny.