Fix double unmutes on `!aegis`
When generating a list of all users to unmute/unban with `!aegis`, we
combine the values in the `aegisCache`, an object that maps nuked
phrases to affected users, and the `singleUserAegisCache`, an object
that maps affected users to nuked phrases. This always results in
duplicate entries. Simply omitting the users in the `aegisCache` fixes
the issue.

The `singleUserAegisCache` is more complete because it includes users
who typed the banned phrase after the nuke was issued.