Fix error where URLs without schemes don't cache
Due to a quirk with how the link regex is built, URLs with a scheme are
matched case-insensitively, but URLs without a scheme are matched
case-sensitively. This quirk is resolved by exporting the final regex
with `i`.

This is important in regard to getLinks, which creates a new RegExp
object from the original with `g`. This approach doesn't merge the flags
in the new object with the old, however, but overwrites the old. As a
result, getLinks no longer detected URLs without a scheme because all
TLDs in the link regex are uppercase.

Maintaining `i` for the RegExp object in getLinks fixes the issue.